Introduce log.h to decouple serverLog to server.h.
